キーフレームの時間をTimeに変更

+++ NiVE3.Plugin/Property/Types/ColorPropertyType.cs
@@ -40,12 +40,12 @@ namespace NiVE3.Plugin.Property.Types
             switch (keyFrames[baseKeyFrameIndex].InterpolationType)
             {
                 case InterpolationType.Linear:
-                    return Interpolation.Linear((Vector4)keyFrame1.Value!, (Vector4)keyFrame2.Value!, keyFrame1.Time, keyFrame2.Time, (double)time);
+                    return Interpolation.Linear((Vector4)keyFrame1.Value!, (Vector4)keyFrame2.Value!, (double)keyFrame1.Time, (double)keyFrame2.Time, (double)time);
                 case InterpolationType.CatmullRom:
                     {
                         var keyFrame0 = baseKeyFrameIndex > 0 ? keyFrames[baseKeyFrameIndex - 1] : keyFrame1;
                         var keyFrame3 = baseKeyFrameIndex <= keyFrames.Count - 3 ? keyFrames[baseKeyFrameIndex + 2] : keyFrame2;
-                        return Interpolation.CatmullRom((Vector4)keyFrame0.Value!, (Vector4)keyFrame1.Value!, (Vector4)keyFrame2.Value!, (Vector4)keyFrame3.Value!, keyFrame1.Time, keyFrame2.Time, (double)time);
+                        return Interpolation.CatmullRom((Vector4)keyFrame0.Value!, (Vector4)keyFrame1.Value!, (Vector4)keyFrame2.Value!, (Vector4)keyFrame3.Value!, (double)keyFrame1.Time, (double)keyFrame2.Time, (double)time);
                     }
                 default:
                     return keyFrame1.Value;
